refactor: record structure
 moving api and models to related record folder
