+++ invenio_records_marc21/records/__init__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2021 Graz University of Technology.
+#
+# Invenio-Records-Marc21 is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+
+"""Marc21 Records module."""
+
+from .api import Marc21Draft, Marc21Record
+from .models import DraftMetadata, RecordMetadata
+
+__all__ = (
+    "Marc21Draft",
+    "Marc21Record",
+    "DraftMetadata",
+    "RecordMetadata",
+)
