+++ invenio_records_marc21/services/config.py
@@ -12,8 +12,8 @@ from invenio_drafts_resources.services.records import RecordDraftServiceConfig
 from invenio_records_resources.services.records.components import MetadataComponent
 from invenio_records_resources.services.records.search import terms_filter
 
-from ..api import Marc21Draft, Marc21Record
-from .components import AccessComponent
+from invenio_records_marc21.records import Marc21Draft, Marc21Record
+
 from .permissions import Marc21RecordPermissionPolicy
 from .schemas import Marc21RecordSchema, MetadataSchema
